Moved quick filters to filter editor
