+++ src-ui/src/app/components/document-list/document-list.component.html
@@ -45,15 +45,8 @@
 
   <div class="btn-group ml-2">
 
-    <button type="button" class="btn btn-sm" [ngClass]="isFiltered ? 'btn-primary' : 'btn-outline-primary'" (click)="showFilter=!showFilter">
-      <svg class="toolbaricon" fill="currentColor">
-        <use xlink:href="assets/bootstrap-icons.svg#funnel" />
-      </svg>
-      {{ showFilter ? 'Hide' : 'Show' }} Filter Editor
-    </button>
-
     <div class="btn-group" ngbDropdown role="group">
-      <button class="btn btn-sm btn-outline-primary dropdown-toggle-split" ngbDropdownToggle></button>
+      <button class="btn btn-sm btn-outline-primary dropdown-toggle" ngbDropdownToggle>Saved Views</button>
       <div class="dropdown-menu" ngbDropdownMenu>
         <ng-container *ngIf="!list.savedViewId">
           <button ngbDropdownItem *ngFor="let config of savedViewConfigService.getConfigs()" (click)="loadViewConfig(config)">{{config.title}}</button>
@@ -69,82 +62,8 @@
 
 </app-page-header>
 
-<div class="row pb-1 mb-3 align-items-right">
-  <div class="btn-toolbar col-auto">
-    <span class="text-muted mt-1 mr-2">Quick Filters:</span>
-    <div class="btn-group ml-2" ngbDropdown role="group">
-      <button class="btn btn-outline-primary btn-sm" id="dropdownTags" ngbDropdownToggle>Tags</button>
-      <div class="dropdown-menu quick-filter" ngbDropdownMenu aria-labelledby="dropdownTags">
-        <div class="list-group list-group-flush">
-          <input class="list-group-item form-control" type="text" [(ngModel)]="filterTagsText" placeholder="Filter tags">
-          <ng-container *ngIf="(tags | filter: filterTagsText).length > 0">
-            <button class="list-group-item list-group-item-action d-flex align-items-center" role="menuitem" *ngFor="let tag of tags | filter: filterTagsText; let i = index" (click)="toggleFilterByTag(tag.id)">
-              <div class="selected-icon mr-1">
-                <svg *ngIf="currentViewIncludesTag(tag.id)" width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-check" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path fill-rule="evenodd" d="M10.97 4.97a.75.75 0 0 1 1.071 1.05l-3.992 4.99a.75.75 0 0 1-1.08.02L4.324 8.384a.75.75 0 1 1 1.06-1.06l2.094 2.093 3.473-4.425a.236.236 0 0 1 .02-.022z"/>
-                </svg>
-              </div>
-              <div class="mr-1"><app-tag [tag]="tag" [clickable]="true" linkTitle="Filter by tag"></app-tag></div>
-              <div class="badge bg-primary text-light rounded-pill ml-auto">{{tag.document_count}}</div>
-            </button>
-          </ng-container>
-        </div>
-      </div>
-    </div>
-
-    <div class="btn-group ml-2" ngbDropdown role="group">
-      <button class="btn btn-outline-primary btn-sm" id="dropdownCorrespondents" ngbDropdownToggle>Correspondents</button>
-      <div class="dropdown-menu quick-filter" ngbDropdownMenu aria-labelledby="dropdownCorrespondents">
-        <div class="list-group list-group-flush">
-          <input class="list-group-item form-control" type="text" [(ngModel)]="filterCorrespondentsText" placeholder="Filter correspondents">
-          <ng-container *ngIf="(correspondents | filter: filterCorrespondentsText).length > 0">
-            <button class="list-group-item list-group-item-action d-flex align-items-center" role="menuitem" *ngFor="let correspondent of correspondents | filter: filterCorrespondentsText; let i = index" (click)="toggleFilterByCorrespondent(correspondent.id)">
-              <div class="selected-icon mr-1">
-                <svg *ngIf="currentViewIncludesCorrespondent(correspondent.id)" width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-check" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path fill-rule="evenodd" d="M10.97 4.97a.75.75 0 0 1 1.071 1.05l-3.992 4.99a.75.75 0 0 1-1.08.02L4.324 8.384a.75.75 0 1 1 1.06-1.06l2.094 2.093 3.473-4.425a.236.236 0 0 1 .02-.022z"/>
-                </svg>
-              </div>
-              <div class="mr-1">{{correspondent.name}}</div>
-              <div class="badge bg-primary text-light rounded-pill ml-auto">{{correspondent.document_count}}</div>
-            </button>
-          </ng-container>
-        </div>
-      </div>
-    </div>
-
-    <div class="btn-group ml-2" ngbDropdown role="group">
-      <button class="btn btn-outline-primary btn-sm" id="dropdownDocumentTypes" ngbDropdownToggle>Document Types</button>
-      <div class="dropdown-menu quick-filter" ngbDropdownMenu aria-labelledby="dropdownDocumentTypes">
-        <div class="list-group list-group-flush">
-          <input class="list-group-item form-control" type="text" [(ngModel)]="filterDocumentTypesText" placeholder="Filter document types">
-          <ng-container *ngIf="(documentTypes | filter: filterDocumentTypesText).length > 0">
-            <button class="list-group-item list-group-item-action d-flex align-items-center" role="menuitem" *ngFor="let documentType of documentTypes | filter: filterDocumentTypesText; let i = index" (click)="toggleFilterByDocumentType(documentType.id)">
-              <div class="selected-icon mr-1">
-                <svg *ngIf="currentViewIncludesDocumentType(documentType.id)" width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-check" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-                  <path fill-rule="evenodd" d="M10.97 4.97a.75.75 0 0 1 1.071 1.05l-3.992 4.99a.75.75 0 0 1-1.08.02L4.324 8.384a.75.75 0 1 1 1.06-1.06l2.094 2.093 3.473-4.425a.236.236 0 0 1 .02-.022z"/>
-                </svg>
-              </div>
-              <div class="mr-1">{{documentType.name}}</div>
-              <div class="badge bg-primary text-light rounded-pill ml-auto">{{documentType.document_count}}</div>
-            </button>
-          </ng-container>
-        </div>
-      </div>
-    </div>
-
-    <button class="btn-link border-0 bg-transparent ml-3 text-muted" *ngIf="currentViewIncludesQuickFilter()" (click)="filterEditor.clearClicked()">
-      <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-x" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-        <path fill-rule="evenodd" d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"/>
-      </svg>
-      Clear
-    </button>
-
-  </div>
-</div>
-
-<div class="card w-100 mb-3" [hidden]="!showFilter">
+<div class="card w-100 mb-3">
   <div class="card-body">
-    <h5 class="card-title">Filter Editor</h5>
     <app-filter-editor [(filterRules)]="filterRules" (apply)="applyFilterRules()" (clear)="clearFilterRules()"></app-filter-editor>
   </div>
 </div>
