--- src-ui/src/app/components/document-list/document-list.component.scss
@@ -1,10 +1,0 @@
-.quick-filter {
-  min-width: 250px;
-  max-height: 400px;
-  overflow-y: scroll;
-  
-  .selected-icon {
-    min-width: 1em;
-    min-height: 1em;
-  }
-}
